Set cookie within frontend
The domain on the cookie set via the intialiser function is scoped to the domain on which the function runs.

Digitalocean has no support for custom function names because it is bobbins.

We have no guarantee that the signup function will have access to that cookie; there are no docs on whether all functions are on the same domain, nor whether functions have access to cookies at all.

Thus, we make the webapp govern this stuff because why the fuck not